clear up the code.
